fix(lens): move the Code-tab lazy/Suspense boundary into CodeTab's own module
The heavy implementation now lives in ./CodeTabImpl; ./CodeTab is a thin lazy() + Suspense wrapper that owns the chunk-splitting and fallback. RightPanel statically imports CodeTab and only gates first-load via codeEverShown, so the router is a pure dispatcher again and the import block is contiguous once more (absorbing hickey-4).

Agreed by the lowy ⇄ hickey lens debate (finding hickey-3, raised by hickey). Not pushed or merged.
